excel_service/requirements.txt
fastapi
uvicorn
openpyxl
psycopg2-binary
sqlalchemy
python-multipart
python-dotenv